Got rid of the dbreader abstractions. Also fixed a bug, where the name of the property in sorting needed to match the case of the property name. Now it is case-insensitive.
