+++ PetFamily.Backend/src/PetFamily.Application/PetsManagement/Pets/Commands/AddPet/AddPetHandler.cs
@@ -1,10 +1,12 @@
 ﻿using CSharpFunctionalExtensions;
+using Dapper;
 using Microsoft.Extensions.Logging;
 using PetFamily.Application.PetsManagement.Volunteers.Interfaces;
 using PetFamily.Application.Shared.Abstractions;
 using PetFamily.Application.Shared.DTOs;
-using PetFamily.Application.SpeciesManagement.Interfaces;
-using PetFamily.Application.SpeciesManagement.Queries.GetBreeds;
+using PetFamily.Application.Shared.Interfaces;
+using PetFamily.Application.SpeciesManagement.DTOs;
+using PetFamily.Domain.Helpers;
 using PetFamily.Domain.PetsManagement.Entities;
 using PetFamily.Domain.PetsManagement.ValueObjects.Pets;
 using PetFamily.Domain.PetsManagement.ValueObjects.Volunteers;
@@ -12,6 +14,7 @@ using PetFamily.Domain.Shared;
 using PetFamily.Domain.Shared.Primitives;
 using PetFamily.Domain.Shared.ValueObjects;
 using PetFamily.Domain.SpeciesManagement.ValueObjects;
+using System.Text;
 
 namespace PetFamily.Application.PetsManagement.Pets.Commands.AddPet;
 
@@ -19,18 +22,18 @@ public class AddPetHandler
     : ICommandHandler<PetId, AddPetCommand>
 {
     private readonly IVolunteerAggregateRepository _volunteerRepository;
-    private readonly ISpeciesAggregateDBReader _speciesDBReader;
+    private readonly IDBConnectionFactory _dBConnectionFactory;
     private readonly AddPetCommandValidator _validator;
     private readonly ILogger<AddPetHandler> _logger;
 
     public AddPetHandler(
         IVolunteerAggregateRepository volunteerRepository,
-        ISpeciesAggregateDBReader speciesDBReader,
+        IDBConnectionFactory dBConnectionFactory,
         AddPetCommandValidator validator,
         ILogger<AddPetHandler> logger)
     {
         _volunteerRepository = volunteerRepository;
-        _speciesDBReader = speciesDBReader;
+        _dBConnectionFactory = dBConnectionFactory;
         _validator = validator;
         _logger = logger;
     }
@@ -58,7 +61,7 @@ public class AddPetHandler
 
         // validate breed
         var breedId = BreedId.Create(command.Pet.BreedId);
-        var breedResult = await _speciesDBReader.GetBreedByIdAsync(breedId.Value, cancellationToken);
+        var breedResult = await GetBreedByIdAsync(breedId.Value, cancellationToken);
         if (breedResult.IsFailure)
             return breedResult.Error;
 
@@ -115,5 +118,31 @@ public class AddPetHandler
             return result.Error;
 
         return petId;
+    }
+
+    public async Task<Result<BreedDTO, ErrorList>> GetBreedByIdAsync(
+        Guid BreedId,
+        CancellationToken cancellationToken = default)
+    {
+        using var connection = _dBConnectionFactory.Create();
+
+        var parameters = new DynamicParameters();
+        parameters.Add("@id", BreedId);
+
+        var sql = new StringBuilder(
+            """
+            SELECT id, name, species_id
+            FROM Breeds
+            WHERE id = @id
+            LIMIT 1
+            """
+        );
+
+        var entity = await connection.QueryFirstAsync<BreedDTO>(sql.ToString(), parameters);
+
+        if (entity == null)
+            return ErrorHelper.General.NotFound(BreedId).ToErrorList();
+
+        return Result.Success<BreedDTO, ErrorList>(entity);
     }
 }
