+++ PetFamily.Backend/src/PetFamily.Application/PetsManagement/Volunteers/Queries/GetVolunteers/GetVolunteersHandler.cs
@@ -1,24 +1,31 @@
 ﻿using CSharpFunctionalExtensions;
+using Dapper;
+using Microsoft.Extensions.Logging;
 using PetFamily.Application.PetsManagement.Volunteers.DTOs;
-using PetFamily.Application.PetsManagement.Volunteers.Interfaces;
 using PetFamily.Application.Shared.Abstractions;
 using PetFamily.Application.Shared.DTOs;
+using PetFamily.Application.Shared.Helpers;
+using PetFamily.Application.Shared.Interfaces;
 using PetFamily.Domain.Shared;
+using System.Text;
 
 namespace PetFamily.Application.PetsManagement.Volunteers.Queries.GetVolunteers;
 
 public class GetVolunteersHandler
     : IQueryHandler<DataListPage<VolunteerDTO>, GetVolunteersQuery>
 {
-    private readonly IVolunteerAggregateDBReader _dbReader;
+    private readonly IDBConnectionFactory _dBConnectionFactory;
     private readonly GetVolunteersQueryValidator _validator;
+    private readonly ILogger<GetVolunteersHandler> _logger;
 
     public GetVolunteersHandler(
-        IVolunteerAggregateDBReader dbReader,
-        GetVolunteersQueryValidator validator)
+        IDBConnectionFactory dBConnectionFactory,
+        GetVolunteersQueryValidator validator,
+        ILogger<GetVolunteersHandler> logger)
     {
-        _dbReader = dbReader;
+        _dBConnectionFactory = dBConnectionFactory;
         _validator = validator;
+        _logger = logger;
     }
 
     public async Task<Result<DataListPage<VolunteerDTO>, ErrorList>> HandleAsync(
@@ -35,6 +42,42 @@ public class GetVolunteersHandler
             return new ErrorList(errors);
         }
 
-        return await _dbReader.GetAsync(query, cancellationToken);
+        using var connection = _dBConnectionFactory.Create();
+
+        var totalCount = await connection.ExecuteScalarAsync<int>("""
+            SELECT Count(*)
+            FROM Volunteers
+            WHERE is_deleted = false
+            """);
+
+        var parameters = new DynamicParameters();
+        parameters.Add("@offset", (query.CurrentPage - 1) * query.PageSize);
+        parameters.Add("@limit", query.PageSize);
+
+        var sql = new StringBuilder(
+            """
+            SELECT id, first_name, last_name, father_name, email, description, experience_years, phone, requisites, social_networks, is_deleted
+            FROM Volunteers
+            WHERE is_deleted = false
+            """);
+
+        if (query.Sort.Any())
+            DapperSQLHelper.ApplySorting(sql, query.Sort);
+
+        DapperSQLHelper.ApplyPagination(sql, parameters, query.CurrentPage, query.PageSize);
+
+        _logger.LogInformation("Dapper SQL: {sql}", sql.ToString());
+
+        var entities = await connection.QueryAsync<VolunteerDTO>(sql.ToString(), parameters);
+
+        var result = new DataListPage<VolunteerDTO>
+        {
+            TotalCount = totalCount,
+            PageNumber = query.CurrentPage,
+            PageSize = query.PageSize,
+            Data = entities
+        };
+
+        return Result.Success<DataListPage<VolunteerDTO>, ErrorList>(result);
     }
 }
