+++ PetFamily.Backend/src/PetFamily.Application/SpeciesManagement/Commands/DeleteBreed/DeleteBreedHandler.cs
@@ -1,9 +1,12 @@
 ﻿using CSharpFunctionalExtensions;
+using Dapper;
 using Microsoft.Extensions.Logging;
 using PetFamily.Application.Shared.Abstractions;
+using PetFamily.Application.Shared.Interfaces;
 using PetFamily.Application.SpeciesManagement.Interfaces;
 using PetFamily.Domain.Shared;
 using PetFamily.Domain.SpeciesManagement.ValueObjects;
+using System.Text;
 
 namespace PetFamily.Application.SpeciesManagement.Commands.DeleteBreed;
 
@@ -11,18 +14,18 @@ public class DeleteBreedHandler
     : ICommandHandler<BreedId, DeleteBreedCommand>
 {
     private readonly ISpeciesAggregateRepository _speciesRepository;
-    private readonly ISpeciesAggregateDBReader _speciesDBReader;
+    private readonly IDBConnectionFactory _dBConnectionFactory;
     private readonly DeleteBreedCommandValidator _validator;
     private readonly ILogger<DeleteBreedHandler> _logger;
 
     public DeleteBreedHandler(
         ISpeciesAggregateRepository speciesRepository,
-        ISpeciesAggregateDBReader speciesDBReader,
+        IDBConnectionFactory dBConnectionFactory,
         DeleteBreedCommandValidator validator,
         ILogger<DeleteBreedHandler> logger)
     {
         _speciesRepository = speciesRepository;
-        _speciesDBReader = speciesDBReader;
+        _dBConnectionFactory = dBConnectionFactory;
         _validator = validator;
         _logger = logger;
     }
@@ -49,7 +52,7 @@ public class DeleteBreedHandler
 
         // check if there are any pet of this breed
         var breedId = BreedId.Create(command.BreedId);
-        var checkPetsResult = await _speciesDBReader.ArePetsWithBreedIdNotExistAsync(breedId.Value, cancellationToken);
+        var checkPetsResult = await ArePetsWithBreedIdNotExistAsync(breedId.Value, cancellationToken);
         if (checkPetsResult.IsFailure)  // if error - pets exist
             return checkPetsResult.Error;
 
@@ -60,5 +63,33 @@ public class DeleteBreedHandler
         _logger.LogInformation("Breed with id {Id} deleted", breedId.Value);
 
         return breedId;
+    }
+
+    public async Task<UnitResult<ErrorList>> ArePetsWithBreedIdNotExistAsync(
+        Guid BreedId,
+        CancellationToken cancellationToken = default)
+    {
+        using var connection = _dBConnectionFactory.Create();
+
+        var parameters = new DynamicParameters();
+        parameters.Add("@breedId", BreedId);
+
+        var sql = new StringBuilder(
+            """
+            SELECT id
+            FROM Pets
+            WHERE breed_id = @breedId
+            LIMIT 1
+            """
+        );
+
+        var result = await connection.QueryAsync<Guid>(sql.ToString(), parameters);
+        if (result is not null && result.Any())
+            return Error.Conflict(
+                "relation.exist",
+                $"Cannot delete breed {BreedId}. It has related pet: {result.First()}")
+                .ToErrorList();
+
+        return UnitResult.Success<ErrorList>();
     }
 }
